+++ lib/index.ts
@@ -4,4 +4,6 @@ export * as Ajax from './interface';
 
 export { default as AjaxBase } from './base';
 
+export { isFormData } from './utils/form';
+
 export default ajax;