+++ lib/index.ts
@@ -6,4 +6,6 @@ export { default as AjaxBase } from './base';
 
 export { isFormData } from './utils/form';
 
+export { promisify } from './utils/promise';
+
 export default ajax;